+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - The build process now automatically detects license files and copies them to `<package-prefix>/share/licenses/<package-name>`.
 - The `link` command now also shows warnings when `--force` is enabled.
 - The install tree now has color coding, based on installed, prebuild and build state.
+- Package deprecation checks now ignore deprecations that are more than a year in the future.
 
 ### Fixed
 - Fix error messages not representing the error correctly.